added investigation overview component to overview mode
